+++ argilla/src/argilla/datasets/_io/__init__.py
@@ -12,7 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from argilla.datasets._export.card._dataset_card import ArgillaDatasetCard
-from argilla.datasets._export.card._parser import size_categories_parser
-
-__all__ = ["ArgillaDatasetCard", "size_categories_parser"]
+from argilla.datasets._io._disk import DiskImportExportMixin  # noqa
+from argilla.datasets._io._hub import HubImportExportMixin  # noqa
