+++ argilla/src/argilla/datasets/_io/card/__init__.py
@@ -12,5 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from argilla.datasets._export._disk import DiskImportExportMixin  # noqa
-from argilla.datasets._export._hub import HubImportExportMixin  # noqa
+from argilla.datasets._io.card._dataset_card import ArgillaDatasetCard
+from argilla.datasets._io.card._parser import size_categories_parser
+
+__all__ = ["ArgillaDatasetCard", "size_categories_parser"]
